+++ packages/drawer/src/views/GestureHandler.android.tsx
@@ -1,0 +1,1 @@
+export * from './GestureHandlerNative';
