+++ packages/drawer/src/views/GestureHandler.ios.tsx
@@ -1,0 +1,1 @@
+export * from './GestureHandlerNative';
