--- packages/drawer/src/views/GestureHandler.macos.tsx
@@ -1,7 +1,0 @@
-export {
-  PanGestureHandler,
-  TapGestureHandler,
-  GestureHandlerRootView,
-  GestureState,
-} from './GestureHandlerStub';
-export type { PanGestureHandlerGestureEvent } from 'react-native-gesture-handler';
