--- packages/stack/src/views/GestureHandler.macos.tsx
@@ -1,6 +1,0 @@
-export {
-  PanGestureHandler,
-  GestureHandlerRootView,
-  GestureState,
-} from './GestureHandlerStub';
-export type { PanGestureHandlerGestureEvent } from 'react-native-gesture-handler';
