+++ packages/stack/src/views/GestureHandlerNative.tsx
