+++ packages/stack/src/views/MaskedView.android.tsx
@@ -1,0 +1,1 @@
+export { default } from './MaskedViewNative';
