+++ packages/stack/src/views/MaskedView.ios.tsx
@@ -1,0 +1,1 @@
+export { default } from './MaskedViewNative';
