--- packages/stack/src/views/MaskedView.macos.tsx
@@ -1,1 +1,0 @@
-export { default } from './MaskedViewStub';
