--- packages/stack/src/views/MaskedView.windows.tsx
@@ -1,1 +1,0 @@
-export { default } from './MaskedViewStub';
