+++ packages/stack/src/views/MaskedViewNative.tsx
@@ -1,3 +1,6 @@
+/**
+ * The native MaskedView that we explicitly re-export for supported platforms: Android, iOS.
+ */
 import * as React from 'react';
 import { UIManager } from 'react-native';
 
@@ -10,6 +13,8 @@ type Props = React.ComponentProps<MaskedViewType> & {
 let RNCMaskedView: MaskedViewType | undefined;
 
 try {
+  // Add try/catch to support usage even if it's not installed, since it's optional.
+  // Newer versions of Metro will handle it properly.
   RNCMaskedView = require('@react-native-community/masked-view').default;
 } catch (e) {
   // Ignore
