--- packages/stack/src/views/TouchableItem.windows.tsx
@@ -1,3 +1,0 @@
-import { TouchableOpacity } from 'react-native';
-
-export const TouchableItem = (TouchableOpacity as any) as typeof import('./TouchableItem.native').TouchableItem;
